VMA permissions use the page table flags.
Also add add_stack(), no env or arg copying yet.

+++ include/sbunix/mm/vmm.h
@@ -15,7 +15,7 @@ struct mm_struct *mm_create_user(uint64_t start_code, uint64_t end_code,
 struct mm_struct *mm_create(void);
 void              mm_destroy(struct mm_struct *mm);
 int               mmap_area(struct mm_struct *mm, struct file *filep,
-                            off_t fstart, size_t fsize, ulong prot,
+                            off_t fstart, size_t fsize, uint64_t prot,
                             uint64_t vm_start, uint64_t vm_end);
 uint64_t          mm_do_sys_sbrk(struct mm_struct *mm, uint64_t newbrk);
 
@@ -29,7 +29,7 @@ void            vma_destroy_all(struct mm_struct *mm);
 struct vm_area *vma_find_region(struct vm_area *vma, uint64_t addr, size_t size);
 
 /* onfault's */
-uint64_t onfault_mmap_file(struct vm_area *vma, uint64_t addr);
-uint64_t onfault_mmap_anon(struct vm_area *vma, uint64_t addr);
+int onfault_mmap_file(struct vm_area *vma, uint64_t addr);
+int onfault_mmap_anon(struct vm_area *vma, uint64_t addr);
 
 #endif
